requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
DateTime==5.5
et_xmlfile==2.0.0
fonttools==4.59.0
fuzz==0.1.1
idna==3.10
jellyfish==1.2.0
joblib==1.5.1
kiwisolver==1.4.8
line_profiler==5.0.0
matplotlib==3.10.3
numpy==2.3.3
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
RapidFuzz==3.13.0
recordlinkage==0.16
requests==2.32.5
requests-oauthlib==2.0.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
survey==5.4.2
thefuzz==0.22.1
threadpoolctl==3.6.0
tweepy==4.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zope.interface==7.2
